test split dictation budget coverage
